Fix issues with series names that have spaces
Since we are using the series names as class names of the plotted
circles, used for hover effects, when the series name has a space,
the circles obtain multiple classes.

E.g. series has the name "series a",
the relevant data points (svg circles) will have classes "series"
and "a".

Hence when trying to select all of them we cannot just select using
the series name "series a", we must replace the spaces with a ".",
so the selection ends up being ".series.a".

This commit implements this behavior to allow flexibility to series names.